reorganize layout components as atomic
