--- packages/ui/src/components/layout/index.ts
@@ -1,4 +1,0 @@
-export * from './NavBar'
-export * from './Paper'
-export * from './Screen'
-export * from './ScreenSection'
