+++ packages/ui/src/components/molecules/index.ts
@@ -1,2 +1,4 @@
 export * from './Blob'
+export * from './NavLink'
 export * from './Pill'
+export * from './ScreenSection'
