+++ packages/ui/src/components/organisms/index.ts
@@ -1,3 +1,4 @@
 export * from './Event'
+export * from './NavBar'
 export * from './PillGroup'
 export * from './PillList'
