+++ packages/ui/src/index.ts
@@ -1,5 +1,4 @@
 export * from './components/cards'
-export * from './components/layout'
 export * from './components/molecules'
 export * from './components/organisms'
 export * from './components/templates'
